feat: 0101_it5

+++ labwork-5/project/src/main/java/task4/User/UserRepository.java
@@ -99,7 +99,7 @@ public class UserRepository extends User implements IRepository <User>{
                 BankAccount bankAccount = new BankAccount(
                         rs.getInt("id"),
                         rs.getString("contributionName"),
-                        rs.getFloat("balance"),
+                        rs.getBigDecimal("balance"),
                         rs.getInt("userId"),
                         rs.getInt("pin"),
                         rs.getString("bankName")
